add new file marvell10g_ptp
