+++ drivers/net/phy/marvell10g.c
@@ -29,15 +29,11 @@
 #include <linux/phy.h>
 #include <linux/sfp.h>
 #include <linux/firmware.h>
-#include <linux/ptp_clock_kernel.h>
-#include <linux/spinlock.h>
 
 #define MV_PHY_ALASKA_NBT_QUIRK_MASK	0xfffffffe
 #define MV_PHY_ALASKA_NBT_QUIRK_REV	(MARVELL_PHY_ID_88X3310 | 0xa)
 
 #define MV_FIRMWARE_HEADER_SIZE		32
-
-#define MV_EXTTS_PERIOD_MS		95
 
 enum {
 	MV_PMA_FW_VER0		= 0xc011,
@@ -92,38 +88,12 @@ enum {
 	MV_AN_STAT1000		= 0x8001, /* 1000base-T status register */
 
 	/* Vendor2 MMD registers */
-	MV_V2_MODE_CFG		= 0xf000,
-	MV_V2_MODE_CFG_M_UNIT_PWRUP = BIT(12),
 	MV_V2_PORT_CTRL		= 0xf001,
 	MV_V2_PORT_CTRL_SWRST	= BIT(15),
 	MV_V2_PORT_CTRL_PWRDOWN = BIT(11),
 	MV_V2_PORT_MAC_TYPE_MASK = 0x7,
 	MV_V2_PORT_MAC_TYPE_XFI_SGMII_AUTONEG = 0x4,
 	MV_V2_PORT_MAC_TYPE_RATE_MATCH = 0x6,
-
-	/* Vendor2 MMD PTP registers */
-	MV_V2_INDIRECT_READ_ADDR	= 0x97fd,
-	MV_V2_INDIRECT_READ_DATA_LOW	= 0x97fe,
-	MV_V2_INDIRECT_READ_DATA_HIGH	= 0x97ff,
-
-	MV_V2_PTP_TOD_LOAD_NSEC_FRAC	= 0xbc2a,
-	MV_V2_PTP_TOD_LOAD_NSEC		= 0xbc2c,
-	MV_V2_PTP_TOD_LOAD_SEC_LOW	= 0xbc2e,
-	MV_V2_PTP_TOD_LOAD_SEC_HIGH	= 0xbc30,
-	MV_V2_PTP_TOD_CAP0_NSEC_FRAC	= 0xbc32,
-	MV_V2_PTP_TOD_CAP0_NSEC		= 0xbc34,
-	MV_V2_PTP_TOD_CAP0_SEC_LOW	= 0xbc36,
-	MV_V2_PTP_TOD_CAP0_SEC_HIGH	= 0xbc38,
-
-	MV_V2_PTP_TOD_CAP_CFG		= 0xbc42,
-	MV_V2_PTP_TOD_CAP_CFG_VAL0	= BIT(0),
-	MV_V2_PTP_TOD_CAP_CFG_VAL1	= BIT(1),
-	MV_V2_PTP_TOD_FUNC_CFG		= 0xbc46,
-	MV_V2_PTP_TOD_FUNC_CFG_TRIG	= BIT(28),
-	MV_V2_PTP_TOD_FUNC_CFG_UPDATE	= 0,
-	MV_V2_PTP_TOD_FUNC_CFG_INCR	= BIT(30),
-	MV_V2_PTP_TOD_FUNC_CFG_DECR	= BIT(31),
-	MV_V2_PTP_TOD_FUNC_CFG_CAPTURE	= BIT(31) | BIT(30),
 
 	/* Temperature control/read registers (88X3310 only) */
 	MV_V2_TEMP_CTRL		= 0xf08a,
@@ -134,57 +104,7 @@ enum {
 	MV_V2_TEMP_UNKNOWN	= 0x9600, /* unknown function */
 };
 
-static int mv3310_adjfine(struct ptp_clock_info *ptp, long scaled_ppm)
-{
-	return -EOPNOTSUPP;
-}
-
-static int mv3310_adjphase(struct ptp_clock_info *ptp, s32 phase)
-{
-	return -EOPNOTSUPP;
-}
-
-static int mv3310_getcrosststamp(struct ptp_clock_info *ptp,
-				 struct system_device_crosststamp *cts)
-{
-	return -EOPNOTSUPP;
-}
-
-static int mv3310_verify(struct ptp_clock_info *ptp, unsigned int pin,
-			 enum ptp_pin_function func, unsigned int chan)
-{
-	return -EOPNOTSUPP;
-}
-
-static int mv3310_adjtime(struct ptp_clock_info *ptp, s64 delta);
-static int mv3310_gettimex64(struct ptp_clock_info *ptp, struct timespec64 *ts,
-			     struct ptp_system_timestamp *sts);
-static int mv3310_settime64(struct ptp_clock_info *p,
-			    const struct timespec64 *ts);
-static int mv3310_enable(struct ptp_clock_info *ptp,
-			 struct ptp_clock_request *request, int on);
-static long mv3310_do_aux_work(struct ptp_clock_info *ptp);
-
-static struct ptp_clock_info mv3310_ptp_clock_info = {
-	.owner = THIS_MODULE,
-	.name = "mv10g-phy-phc",
-	.max_adj = 0,
-	.n_alarm = 0,
-	.n_ext_ts = 1,
-	.n_per_out = 0,
-	.n_pins = 0,
-	.pps = 0,
-	.pin_config = NULL,
-	.adjfine = mv3310_adjfine,
-	.adjphase = mv3310_adjphase,
-	.adjtime = mv3310_adjtime,
-	.gettimex64 = mv3310_gettimex64,
-	.getcrosststamp = mv3310_getcrosststamp,
-	.settime64 = mv3310_settime64,
-	.enable = mv3310_enable,
-	.verify = mv3310_verify,
-	.do_aux_work = mv3310_do_aux_work,
-};
+struct mv3310_ptp_priv;
 
 struct mv3310_priv {
 	u32 firmware_ver;
@@ -194,12 +114,13 @@ struct mv3310_priv {
 	struct device *hwmon_dev;
 	char *hwmon_name;
 
-	struct phy_device *phydev;
-	struct ptp_clock *ptp_clock;
-	struct ptp_clock_info ptp_clock_info;
-	spinlock_t ptp_lock;
-	bool extts_enabled;
+	struct mv3310_ptp_priv *ptp_priv;
 };
+
+struct mv3310_ptp_priv *mv3310_ptp_probe(struct phy_device *phydev);
+int mv3310_ptp_power_up(struct phy_device *phydev);
+int mv3310_ptp_power_down(struct phy_device *phydev);
+int mv3310_ptp_resume(struct phy_device *phydev);
 
 #ifdef CONFIG_HWMON
 static umode_t mv3310_hwmon_is_visible(const void *data,
@@ -351,298 +272,9 @@ static int mv3310_hwmon_probe(struct phy_device *phydev)
 }
 #endif
 
-static int mv3310_read_ptp_reg(struct phy_device *phydev, u32 regnum, u32 *regval)
-{
-	int ret;
-
-	/* Read register address */
-	ret = phy_read_mmd(phydev, MDIO_MMD_VEND2, regnum);
-	if (ret < 0)
-		return ret;
-
-	/* Read that Indirect_read_address gives requested address */
-	ret = phy_read_mmd(phydev, MDIO_MMD_VEND2, MV_V2_INDIRECT_READ_ADDR);
-	if (ret < 0)
-		return ret;
-	if (ret != regnum) {
-		pr_err("Indirect read address mismatch: %04x != %04x\n", ret,
-		       regnum);
-		return -EINVAL;
-	}
-
-	/* Read Indirect_read_data_low provides lower 16-bits (15:0) of data */
-	ret = phy_read_mmd(phydev, MDIO_MMD_VEND2,
-			   MV_V2_INDIRECT_READ_DATA_LOW);
-	if (ret < 0)
-		return ret;
-	*regval = ret & 0xffff;
-
-	/* Read Indirect_read_data_high provides upper 16-bits (31:16) of data */
-	ret = phy_read_mmd(phydev, MDIO_MMD_VEND2,
-			   MV_V2_INDIRECT_READ_DATA_HIGH);
-	if (ret < 0)
-		return ret;
-	*regval += ((ret & 0xffff) << 16);
-
-	return 0;
-}
-
-static int mv3310_write_ptp_reg(struct phy_device *phydev, u32 regnum, u32 regval)
-{
-	int ret;
-
-	ret = phy_write_mmd(phydev, MDIO_MMD_VEND2, regnum, regval);
-	if (ret < 0)
-		return ret;
-
-	ret = phy_write_mmd(phydev, MDIO_MMD_VEND2, regnum + 1, regval >> 16U);
-	if (ret < 0)
-		return ret;
-
-	return 0;
-}
-
-static int mv3310_trigger_ptp_op(struct phy_device *phydev, int op)
-{
-	int ret;
-
-	ret = mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_FUNC_CFG,
-				   MV_V2_PTP_TOD_FUNC_CFG_TRIG | op);
-	if (ret < 0)
-		return ret;
-
-	if (op != MV_V2_PTP_TOD_FUNC_CFG_CAPTURE) {
-		/* Restore capture mode */
-		return mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_FUNC_CFG,
-					    MV_V2_PTP_TOD_FUNC_CFG_CAPTURE);
-	}
-
-	return 0;
-}
-
-static int mv3310_getppstime(struct ptp_clock_info *ptp, struct timespec64 *ts)
-{
-	int ret;
-	u32 nsec_frac = 0, nsec = 0, sec_low = 0, sec_high = 0, cap_cfg = 0;
-
-	struct mv3310_priv *priv =
-		container_of(ptp, struct mv3310_priv, ptp_clock_info);
-	struct phy_device *phydev = priv->phydev;
-
-	/* Check if TOD@pps is available */
-	ret = mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP_CFG, &cap_cfg);
-	if (ret < 0)
-		return ret;
-	if (!(cap_cfg & MV_V2_PTP_TOD_CAP_CFG_VAL0))
-		return -EAGAIN;
-
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_NSEC_FRAC,
-				   &nsec_frac);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_NSEC, &nsec);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_SEC_LOW,
-				   &sec_low);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_SEC_HIGH,
-				   &sec_high);
-
-	if (ret < 0) {
-		pr_err("Failed to read TOD<0> capture: nsec_frac=%u, nsec=%u, sec_low=%u, sec_high=%u\n",
-		       nsec_frac, nsec, sec_low, sec_high);
-		return -EIO;
-	}
-
-	/* Finished reading capture, reset */
-	mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_CAP_CFG, 0);
-
-	if (nsec_frac >= 0x80000000)
-		nsec++;
-	ts->tv_sec = ((u64)sec_high << 32U) | sec_low;
-	ts->tv_nsec = nsec;
-
-	return 0;
-}
-
-static int mv3310_gettimex64(struct ptp_clock_info *ptp, struct timespec64 *ts,
-			     struct ptp_system_timestamp *sts)
-{
-	int ret;
-	unsigned long flags;
-	u32 nsec_frac = 0, nsec = 0, sec_low = 0, sec_high = 0;
-
-	struct mv3310_priv *priv =
-		container_of(ptp, struct mv3310_priv, ptp_clock_info);
-	struct phy_device *phydev = priv->phydev;
-
-	spin_lock_irqsave(&priv->ptp_lock, flags);
-	/* Clear existing TOD Capture Values and trigger new capture.
-	   In the unlikely event that a pulse-in trigger will capture the TOD
-	   to TOD_CAP0 and this CPU trigger will capture it to TOD_CAP1, we are
-	   still reading from TOD_CAP0 as they will be almost equal. */
-	ret = mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_CAP_CFG, 0);
-	if (ret < 0)
-		goto unlock_out;
-
-	ret = mv3310_trigger_ptp_op(phydev, MV_V2_PTP_TOD_FUNC_CFG_CAPTURE);
-	if (ret < 0)
-		goto unlock_out;
-
-	/* Read capture */
-	ptp_read_system_prets(sts);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_NSEC_FRAC,
-				   &nsec_frac);
-	ptp_read_system_postts(sts);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_NSEC, &nsec);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_SEC_LOW,
-				   &sec_low);
-	ret |= mv3310_read_ptp_reg(phydev, MV_V2_PTP_TOD_CAP0_SEC_HIGH,
-				   &sec_high);
-
-	if (ret < 0) {
-		pr_err("Failed to read TOD<0> capture: nsec_frac=%u, nsec=%u, sec_low=%u, sec_high=%u\n",
-		       nsec_frac, nsec, sec_low, sec_high);
-		ret = -EIO;
-		goto unlock_out;
-	}
-
-	/* Finished reading capture, reset */
-	mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_CAP_CFG, 0);
-	spin_unlock_irqrestore(&priv->ptp_lock, flags);
-
-	if (nsec_frac >= 0x80000000)
-		nsec++;
-	ts->tv_sec = ((u64)sec_high << 32U) | sec_low;
-	ts->tv_nsec = nsec;
-
-	return 0;
-
-unlock_out:
-	spin_unlock_irqrestore(&priv->ptp_lock, flags);
-	return ret;
-}
-
-static int mv3310_settime64(struct ptp_clock_info *p,
-			    const struct timespec64 *ts)
-{
-	int ret = 0;
-	unsigned long flags;
-	u32 sec_low = lower_32_bits(ts->tv_sec);
-	u32 sec_high = upper_32_bits(ts->tv_sec) & 0xffff;
-	u32 nsec = lower_32_bits(ts->tv_nsec);
-
-	struct mv3310_priv *priv =
-		container_of(p, struct mv3310_priv, ptp_clock_info);
-	struct phy_device *phydev = priv->phydev;
-
-	/* Load the new timestamp */
-	spin_lock_irqsave(&priv->ptp_lock, flags);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_NSEC_FRAC, 0);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_NSEC, nsec);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_SEC_LOW,
-				    sec_low);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_SEC_HIGH,
-				    sec_high);
-	if (ret < 0) {
-		ret = -EIO;
-		goto unlock_out;
-	}
-
-	/* Trigger update */
-	ret = mv3310_trigger_ptp_op(phydev, MV_V2_PTP_TOD_FUNC_CFG_UPDATE);
-
-unlock_out:
-	spin_unlock_irqrestore(&priv->ptp_lock, flags);
-	return ret;
-}
-
-static int mv3310_adjtime(struct ptp_clock_info *ptp, s64 delta)
-{
-	int ret = 0;
-	unsigned long flags;
-	struct timespec64 ts = ns_to_timespec64(abs(delta));
-
-	struct mv3310_priv *priv =
-		container_of(ptp, struct mv3310_priv, ptp_clock_info);
-	struct phy_device *phydev = priv->phydev;
-
-	if (delta == 0)
-		return 0;
-
-	/* Load the new timestamp */
-	spin_lock_irqsave(&priv->ptp_lock, flags);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_NSEC_FRAC, 0);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_NSEC,
-				    ts.tv_nsec);
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_SEC_LOW,
-				    lower_32_bits(ts.tv_sec));
-	ret |= mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_LOAD_SEC_HIGH,
-				    upper_32_bits(ts.tv_sec) & 0xffff);
-	if (ret < 0) {
-		ret = -EIO;
-		goto unlock_out;
-	}
-
-	/* Trigger update */
-	ret = mv3310_trigger_ptp_op(phydev,
-				    delta < 0 ? MV_V2_PTP_TOD_FUNC_CFG_DECR :
-						MV_V2_PTP_TOD_FUNC_CFG_INCR);
-unlock_out:
-	spin_unlock_irqrestore(&priv->ptp_lock, flags);
-	return ret;
-}
-
-static int mv3310_enable(struct ptp_clock_info *ptp,
-			 struct ptp_clock_request *request, int on)
-{
-	int ret = 0;
-	struct mv3310_priv *priv =
-		container_of(ptp, struct mv3310_priv, ptp_clock_info);
-
-	switch (request->type) {
-	case PTP_CLK_REQ_EXTTS:
-		priv->extts_enabled = on != 0;
-
-		if (priv->extts_enabled)
-			ptp_schedule_worker(priv->ptp_clock, 0);
-		else
-			ptp_cancel_worker_sync(priv->ptp_clock);
-		break;
-
-	default:
-		ret = -EOPNOTSUPP;
-		break;
-	}
-
-	return ret;
-}
-
-static long mv3310_do_aux_work(struct ptp_clock_info *ptp)
-{
-	unsigned long flags;
-	struct mv3310_priv *priv =
-		container_of(ptp, struct mv3310_priv, ptp_clock_info);
-
-	if (priv->extts_enabled) {
-		struct ptp_clock_event event;
-		struct timespec64 ts;
-
-		spin_lock_irqsave(&priv->ptp_lock, flags);
-		if (mv3310_getppstime(ptp, &ts) == 0) {
-			event.type = PTP_CLOCK_EXTTS;
-			event.index = 0; /* We only have one channel */
-			event.timestamp = timespec64_to_ns(&ts);
-			ptp_clock_event(priv->ptp_clock, &event);
-		}
-		spin_unlock_irqrestore(&priv->ptp_lock, flags);
-
-		return msecs_to_jiffies(MV_EXTTS_PERIOD_MS);
-	}
-
-	return msecs_to_jiffies(2000);
-}
-
 static int mv3310_power_down(struct phy_device *phydev)
 {
-	phy_clear_bits_mmd(phydev, MDIO_MMD_VEND2, MV_V2_MODE_CFG,
-			   MV_V2_MODE_CFG_M_UNIT_PWRUP);
+	mv3310_ptp_power_down(phydev);
 	return phy_set_bits_mmd(phydev, MDIO_MMD_VEND2, MV_V2_PORT_CTRL,
 				MV_V2_PORT_CTRL_PWRDOWN);
 }
@@ -658,9 +290,7 @@ static int mv3310_power_up(struct phy_device *phydev)
 	if (ret < 0)
 		return ret;
 
-	/* Enable M unit used for PTP */
-	ret = phy_set_bits_mmd(phydev, MDIO_MMD_VEND2, MV_V2_MODE_CFG,
-			 MV_V2_MODE_CFG_M_UNIT_PWRUP);
+	ret = mv3310_ptp_power_up(phydev);
 	if (ret < 0)
 		return ret;
 
@@ -904,14 +534,7 @@ static int mv3310_probe(struct phy_device *phydev)
 	if (!priv)
 		return -ENOMEM;
 
-	priv->phydev = phydev;
-	priv->ptp_clock_info = mv3310_ptp_clock_info;
-	priv->ptp_clock = ptp_clock_register(&priv->ptp_clock_info, &phydev->mdio.dev);
-	if (IS_ERR(priv->ptp_clock)) {
-		dev_err(&phydev->mdio.dev, "failed to register PTP clock\n");
-		return PTR_ERR(priv->ptp_clock);
-	}
-
+	priv->ptp_priv = mv3310_ptp_probe(phydev);
 	dev_set_drvdata(&phydev->mdio.dev, priv);
 
 	/* Powering down the port when not in use saves about 600mW */
@@ -936,19 +559,6 @@ static int mv3310_suspend(struct phy_device *phydev)
 	return mv3310_power_down(phydev);
 }
 
-static int mv3310_ptp_config(struct phy_device *phydev)
-{
-	int ret;
-
-	/* Enable capture mode */
-	ret = mv3310_write_ptp_reg(phydev, MV_V2_PTP_TOD_FUNC_CFG,
-				   MV_V2_PTP_TOD_FUNC_CFG_CAPTURE);
-	if (ret < 0)
-		return ret;
-
-	return 0;
-}
-
 static int mv3310_resume(struct phy_device *phydev)
 {
 	int ret;
@@ -957,7 +567,7 @@ static int mv3310_resume(struct phy_device *phydev)
 	if (ret)
 		return ret;
 
-	ret = mv3310_ptp_config(phydev);
+	ret = mv3310_ptp_resume(phydev);
 	if (ret)
 		return ret;
